refactor: リクエスト検証とリクエストの取得方法をカプセル化
