+++ KAMAGI/SootControllers/UserController.php
@@ -3,14 +3,14 @@
 namespace KAMAGI\SootControllers;
 
 use KAMAGI\SootUseCases\signUpUseCase;
-use KAMAGI\Response;
+use KAMAGI\SootResources\Response;
 
 /**
  * ユーザーコントローラー
  * 
  * ユーザー関連のリクエストを処理
  */
-class UserController
+class UserController extends BaseController
 {
     private signUpUseCase $signUpUseCase;
 
@@ -26,11 +26,12 @@ class UserController
      */
     public function store(): void
     {
-        $input = json_decode(file_get_contents('php://input'), true);
+        $this->validateMethod('POST');
+        $input = $this->getRequestInput();
 
-        $userId = htmlspecialchars($input['user_id'] ?? '', ENT_QUOTES, 'UTF-8');
-        $name = htmlspecialchars($input['name'] ?? '', ENT_QUOTES, 'UTF-8');
-        $password = htmlspecialchars($input['password'] ?? '', ENT_QUOTES, 'UTF-8');
+        $userId = htmlspecialchars($input['user_id'], ENT_QUOTES, 'UTF-8');
+        $name = htmlspecialchars($input['name'], ENT_QUOTES, 'UTF-8');
+        $password = htmlspecialchars($input['password'], ENT_QUOTES, 'UTF-8');
 
         $result = $this->signUpUseCase->execute($userId, $name, $password);
